Falta so adicionar a imagem
