+++ src/Componentes/InfosUsuario.js
@@ -1,0 +1,14 @@
+import React from "react";
+import Imagem from "../Img/Usuario.png"
+
+export function InfosUsuario() {
+
+    return(
+        <h3>Usuario 1
+            
+        </h3>
+
+        // <img src = {Imagem} alt=""/>
+    )
+
+}
